+++ src/isr.cpp
@@ -1,7 +1,7 @@
 #include "globals.h"
 #include "sid.h"
 #include "lfo.h"
-#include "leds.h"
+#include "ui_leds.h"
 #include "isr.h"
 #include "ui.h"
