+++ src/ui_buttons.cpp
@@ -1,5 +1,5 @@
 #include "globals.h"
-#include "leds.h"
+#include "ui_leds.h"
 #include "preset.h"
 #include "sid.h"
 #include "midi.h"
